DataEnhancement/frontend/requirements.txt
requests==2.32.3
PyJWT==2.10.1
streamlit==1.43.2
streamlit-cookies-controller==0.0.4
streamlit-elements
streamlit-aggrid
hydralit-components